atomic-host-installer: Switch to storing the OSTree repo in /
We hit an issue with having signed commits in the ISO, since
Joliet filenames are severely limited and `$hash.commit` and `$hash.commitmeta`
conflict.

What we want to do long term is change Lorax to use OSTree itself for
Anaconda - that would get us deduplication, and that'd be a big deal
given that we're pushing 900MB now.

This doesn't implement dedup, but it moves us in that direction by
initializing a `bare` mode `/ostree/repo` in the Anaconda environment,
which we pull from.

NOTE NOTE NOTE
--------------
This requires `--rootfs-size=3` and switching both templates to `--add-template`.
Like this:

```
lorax --rootfs-size=3
      --add-template=$(pwd)/fedora-lorax-templates/atomic-installer/lorax-configure-repo.tmpl
      --add-template=$(pwd)/fedora-lorax-templates/atomic-installer/lorax-embed-repo.tmpl
      --add-template-var=ostree_osname=fedora-atomic
      --add-template-var=ostree_ref=fedora-atomic/25/x86_64/docker-host
      --add-template-var=ostree_repo=http://192.168.122.1/fedora-atomic/repo-upstream
```

Signed-off-by: Colin Walters <walters@verbum.org>