atomic: only create new ref if needed
If ostree_install_ref and ostree_update_ref are the same, as is the
case for rawhide, then there is no reason to create a new ref. This
is actually causing failures in rawhide right now.

Having this setup will allow us to have the same configs in rawhide
and branched where we do need separate install/update ref.

Signed-off-by: Dusty Mabe <dusty@dustymabe.com>