lorax-embed-flatpaks.tmpl: Run the flatpak-install under dbus-run-session
Recent versions of Flatpak unconditionally use the Flatpak OCI
authenticator for OCI remotes - even if they don't actually need
authentication. (This simplifies making Flatpak work with registries
that do need authentication, or at least a token.)

But for the OCI authenticator to work, we need a D-Bus session,
so use dbus-run-session to create a temporary one.

Signed-off-by: Owen W. Taylor <otaylor@fishsoup.net>